+++ src/util/creator.js
@@ -1,10 +1,11 @@
-var ctyp = require('./ctyp'),
-		reduce = require('./reduce')
+var reduce = require('./reduce'),
+		text = require('../text'),
+		ENV = require('../env')
 
 module.exports = function creator(constructor) {
 	return function create(defaults) {
 		return function define(selector) {
-			var options = merge({}, defaults),
+			var options = mergeOptions({}, defaults),
 					content = []
 			for (var i=1; i<arguments.length; ++i) argument(arguments[i], options, content)
 			return constructor(selector, options, content)
@@ -12,31 +13,30 @@ module.exports = function creator(constructor) {
 	}
 }
 function argument(arg, options, content) {
-	switch(ctyp(arg)) {
-		case Object:
-			merge(options, arg)
-			break
+	if (arg && arg.constructor === Object) mergeOptions(options, arg)
+	else mergeChildren.call(content, arg)
+}
+function mergeChildren(arg) {
+	if (arg != null) switch(ctyp(arg)) { //eslint-disable-line eqeqeq
 		case Array:
-			flatConcat(content, arg)
-			break
-		case null: case undefined:
+			arg.forEach(mergeChildren, this)
 			break
 		case Number:
-			content.push(''+arg)
+			this.push(text(''+arg))
 			break
-		default:
-			content.push(arg)
+		case String:
+			this.push(text(arg))
 			break
+		default: this.push(arg)
 	}
 }
 /**
- * @param {...Object} tgt - target object
+ * @param {Object} tgt - target object
+ * @param {Object} src - target object
  * @returns {Object} - modified target
  */
-function merge(tgt) {
-	for (var i=1; i<arguments.length; ++i) {
-		if (arguments[i] != null) tgt = reduce(arguments[i], assign, tgt || {}) //eslint-disable-line eqeqeq
-	}
+function mergeOptions(tgt, src) {
+	if (src != null) tgt = reduce(src, assign, tgt || {}) //eslint-disable-line eqeqeq
 	return tgt
 }
 function assign(tgt, val, key) {
@@ -45,14 +45,17 @@ function assign(tgt, val, key) {
 }
 function submerge(tgt, src) {
 	switch(ctyp(src)) {
-		case Array: return flatConcat(tgt || [], src)
+		case Array:
+			if (!tgt) return src.slice()
+			tgt.push.apply(src)
+			return tgt
 		case Object: return reduce(src, assign, tgt || {})
 		case undefined: return tgt
 		default: return src //Number, String, Function, Boolean, null
 	}
 }
-function flatConcat(arr, val) {
-	if (Array.isArray(val)) for (var i=0; i<val.length; ++i) flatConcat(arr, val[i])
-	else arr.push(val)
-	return arr
+function ctyp(t) {
+	return t == null ? t //eslint-disable-line eqeqeq
+		: (t.nodeName && t.nodeType && t.cloneNode) ? ENV.Node
+		: t.constructor || Object
 }